+++ tests/test_interface.py
@@ -1,143 +1,41 @@
 import json
 import os
-from ast import literal_eval
-from base64 import b64decode, b64encode
+from base64 import b64encode
 from datetime import datetime, timezone
-from io import BytesIO
 from pathlib import Path
 from types import SimpleNamespace
 
 import geckodriver_autoinstaller
 import imagehash
 import pytest
-from cairosvg import svg2png
 from flaky import flaky
-from PIL import Image
-from rdkit import Chem, RDConfig
+from rdkit import Chem
 from rdkit import __version__ as rdkit_version
 from rdkit.Chem import AllChem
 from selenium import webdriver
-from selenium.common.exceptions import (
-    NoSuchElementException,
-    StaleElementReferenceException,
-)
+from selenium.common.exceptions import NoSuchElementException
 from selenium.webdriver.common.action_chains import ActionChains
 from selenium.webdriver.common.by import By
 from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.support import expected_conditions as EC
-from selenium.webdriver.support.ui import WebDriverWait
 
 import mols2grid
 from mols2grid.select import register
 from mols2grid.utils import env
 
+from .webdriver_utils import FirefoxDriver
+
 geckodriver_autoinstaller.install()
 pytestmark = pytest.mark.webdriver
 GITHUB_ACTIONS = os.environ.get("GITHUB_ACTIONS")
-
 HEADLESS = True
 PAGE_LOAD_TIMEOUT = 10
-
-
-class selection_available:
-    def __init__(self, is_empty=False):
-        self.empty = is_empty
-
-    def __call__(self, driver):
-        sel = driver.execute_script("return SELECTION.to_dict();")
-        sel = literal_eval(sel)
-        if sel == {} and self.empty:
-            return True
-        elif sel != {} and not self.empty:
-            return sel
-        return False
-
-
-class FirefoxDriver(webdriver.Firefox):
-    def wait_for_img_load(self, max_delay=5, selector="#mols2grid .cell .data-img svg"):
-        return WebDriverWait(self, max_delay).until(
-            EC.presence_of_element_located((By.CSS_SELECTOR, selector))
-        )
-
-    def wait_for_selection(self, is_empty=False, max_delay=3):
-        return WebDriverWait(self, max_delay).until(selection_available(is_empty))
-
-    def wait(self, condition, max_delay=5):
-        return WebDriverWait(
-            self, max_delay, ignored_exceptions=[StaleElementReferenceException]
-        ).until(condition)
-
-    def find_by_id(self, element_id, **kwargs):
-        condition = EC.presence_of_element_located((By.ID, element_id))
-        return self.wait(condition, **kwargs)
-
-    def find_clickable(self, by, selector, **kwargs):
-        condition = EC.element_to_be_clickable((by, selector))
-        return self.wait(condition, **kwargs)
-
-    def trigger_callback(self, selector="#mols2grid .cell .data-img *", pause=0.2):
-        self.wait_for_img_load()
-        el = self.find_clickable(By.CSS_SELECTOR, selector)
-        (ActionChains(self).move_to_element(el).pause(pause).click().perform())
-
-    def find_by_css_selector(self, css_selector, **kwargs):
-        condition = EC.presence_of_element_located((By.CSS_SELECTOR, css_selector))
-        return self.wait(condition, **kwargs)
-
-    def find_by_class_name(self, name, **kwargs):
-        condition = EC.presence_of_element_located((By.CLASS_NAME, name))
-        return self.wait(condition, **kwargs)
-
-    def find_all_by_class_name(self, name, **kwargs):
-        condition = EC.presence_of_all_elements_located((By.CLASS_NAME, name))
-        return self.wait(condition, **kwargs)
-
-    def get_svg_hash(self, *args, **kwargs):
-        im = next(self.get_imgs_from_svgs(*args, **kwargs))
-        return imagehash.average_hash(im, hash_size=16)
-
-    def get_imgs_from_svgs(self, selector="#mols2grid .cell .data-img"):
-        condition = EC.presence_of_all_elements_located((By.CSS_SELECTOR, selector))
-        svgs = self.wait(condition)
-        for svg in svgs:
-            im = svg2png(bytestring=(svg.get_attribute("innerHTML")))
-            yield Image.open(BytesIO(im))
-
-    def get_png_hash(self, selector="#mols2grid .cell .data-img *"):
-        img = self.find_by_css_selector(selector)
-        im = Image.open(BytesIO(b64decode(img.get_attribute("src")[22:])))
-        return imagehash.average_hash(im, hash_size=16)
-
-    def substructure_query(self, smarts):
-        self.find_clickable(By.ID, "searchBtn").click()
-        self.find_clickable(By.ID, "smartsSearch").click()
-        self.find_by_id("searchbar").send_keys(smarts)
-        self.wait_for_img_load()
-
-    def text_search(self, txt):
-        self.find_clickable(By.ID, "searchBtn").click()
-        self.find_clickable(By.ID, "txtSearch").click()
-        self.find_by_id("searchbar").send_keys(txt)
-        self.wait_for_img_load()
 
 
 def determine_scope(fixture_name, config):
     if GITHUB_ACTIONS:
         return "function"
     return "module"
-
-
-@pytest.fixture(scope=determine_scope)
-def driver():
-    options = webdriver.FirefoxOptions()
-    options.headless = True if GITHUB_ACTIONS else HEADLESS
-    driver = FirefoxDriver(options=options)
-    driver.set_page_load_timeout(PAGE_LOAD_TIMEOUT)
-    yield driver
-    driver.quit()
-
-
-sdf_path = f"{RDConfig.RDDocsDir}/Book/data/solubility.test.sdf"
 
 
 def get_grid(df, **kwargs):
@@ -151,19 +49,14 @@ def get_doc(grid, kwargs):
     return "data:text/html;base64,{}".format(html)
 
 
-@pytest.fixture(scope="module")
-def df():
-    return mols2grid.sdf_to_dataframe(sdf_path)[:30]
-
-
-@pytest.fixture(scope="module")
-def mols(df):
-    return df["mol"][:5]
-
-
-@pytest.fixture(scope="module")
-def grid(df):
-    return get_grid(df)
+@pytest.fixture(scope=determine_scope)
+def driver():
+    options = webdriver.FirefoxOptions()
+    options.headless = True if GITHUB_ACTIONS else HEADLESS
+    driver = FirefoxDriver(options=options)
+    driver.set_page_load_timeout(PAGE_LOAD_TIMEOUT)
+    yield driver
+    driver.quit()
 
 
 @pytest.fixture(scope="module")
@@ -805,7 +698,7 @@ def test_custom_header(driver: FirefoxDriver, grid):
     assert val == "2021.09.2"
 
 
-def test_table_template(driver: FirefoxDriver):
+def test_table_template(driver: FirefoxDriver, sdf_path):
     df = mols2grid.sdf_to_dataframe(sdf_path)[:15]
     grid = mols2grid.MolGrid(df, mol_col="mol", prerender=True)
     doc = get_doc(